final commmit before presentation...
wow, this is a monster of a commit.... I had less than 3 hours to finish
this project before I had to present and thus didn't have the time nor
energy to make good clean git commits. As far as functionality, this
commit puts together everything to display a rotating sphere on the
browser using quaternions, however it is held together by glue and duct
tape....

I had to scrap a few features last minute, as around 30 minutes before the
presentation, I broke the app while finishing up a crucial piece of math, and I
began to fear that it would take me too long to diagnose and fix. I
ended up having to fall back on Three.js's built in quaternion methods
instead of calculating it myself so I would have something to present.
My future plans include going back and fixing this issue, as well as
making the app interactive, allowing users to control the rotation of
the sphere and view the changes to the corresponding quaternion in real
time.